refactor(ui-base): move toast and modal_frame down to dbflux_ui_base (#102)
Move toast.rs and modal_frame.rs from dbflux_ui into dbflux_ui_base so
dbflux_ui_document can reach them without an upward dependency on dbflux_ui.
Leave re-export shims at the original crate::ui::components paths so all
41 existing call-sites remain unchanged. Update the modal_frame test to
reflect the new crate layout (CARGO_MANIFEST_DIR now points at ui_base)
and remove the two document/* entries that will move to dbflux_ui_document.
